Add files via upload
This code creates a new figure and adds a subplot to it. It then plots the lamps as circles using the add_patch function and plots the edges of the area using the plot function. The plot is then displayed using the show function. 

First, you will need to install matplotlib ,then, you can use the following code to create a visual representation of the area and the lamps: